+++ .changeset/deduplicate-extension-collection.md
@@ -1,0 +1,5 @@
+---
+'@backstage/frontend-plugin-api': patch
+---
+
+Fixed the duplicate extension error message in `createFrontendModule` to correctly say "Module" instead of "Plugin".
